Add Material "Snackbar", restructure projector
- Since MaterialUI does not have an equivalent to bootstrap alterts,
the toast service and alert service are marked as deprecated for now.
The functionality has been replaced with the snackbar.
