+++ client/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { LoginComponent } from './site/login/login.component';
-import { ProjectorComponent } from './projector/projector.component';
-import { ProjectorContainerComponent } from './projector/projector-container/projector-container.component';
+import { ProjectorComponent } from './projector-container/projector/projector.component';
+import { ProjectorContainerComponent } from './projector-container/projector-container.component';
 import { SiteComponent } from './site/site.component';
 import { StartComponent } from './site/start/start.component';
 import { AgendaComponent } from './site/agenda/agenda.component';
