+++ client/src/app/core/services/alert.service.ts
@@ -2,6 +2,10 @@ import { Injectable } from '@angular/core';
 
 import { Alert, AlertType } from 'app/core/models/alert';
 
+/**TODO Drafted for now. Since the UI is not done yet, this might be replaced or disappear entirely.
+ * Furtermore, Material UI does not support these kinds of alerts
+ */
+
 @Injectable({
     providedIn: 'root'
 })
