+++ client/src/app/core/services/toast.service.ts
@@ -3,6 +3,10 @@ import { Observable, Subject } from 'rxjs';
 
 import { Alert, AlertType } from 'app/core/models/alert';
 
+/**TODO Drafted for now. Since the UI is not done yet, this might be replaced or disappear entirely.
+ * Furtermore, Material UI does not support these kinds of alerts
+ */
+
 @Injectable({
     providedIn: 'root'
 })
