+++ client/src/app/projector-container/projector-container.component.spec.ts
