+++ client/src/app/projector-container/projector/projector.component.html
