+++ client/src/app/projector-container/projector/projector.component.spec.ts
