+++ client/src/app/site/login/login.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Title } from '@angular/platform-browser';
+import { MatSnackBar } from '@angular/material';
 
 import { BaseComponent } from 'app/base.component';
 import { User } from 'app/core/models/user';
 import { AuthService } from 'app/core/services/auth.service';
-import { AlertService } from 'app/core/services/alert.service';
 
 @Component({
     selector: 'app-login',
@@ -22,8 +22,8 @@ export class LoginComponent extends BaseComponent implements OnInit {
     constructor(
         titleService: Title,
         private authService: AuthService,
-        private alertService: AlertService,
-        private router: Router
+        private router: Router,
+        private snackBar: MatSnackBar
     ) {
         super(titleService);
         this.setInfo();
@@ -38,19 +38,23 @@ export class LoginComponent extends BaseComponent implements OnInit {
         this.info = 'Logged in? ' + (this.authService.isLoggedIn ? 'in' : 'out');
     }
 
+    openSnackBar(message: string) {
+        this.snackBar.open(message, 'OK', {
+            duration: 2000
+        });
+    }
+
     //Todo: This serves as a prototype and need enhancement,
     //like saving a "logged in state" and real checking the server
     //if logIn was fine
     formLogin(): void {
         this.authService.login(this.user).subscribe(res => {
             if (res.status === 400) {
-                //TODO, add more errors here, use
-                console.log('error in login');
-
-                //todo alert seems to work differently than toast
-                this.alertService.error('Benutzername oder Passwort war nicht korrekt.');
+                //TODO translate
+                console.log('res: ', res);
+                this.openSnackBar(res.error.detail);
             } else {
-                this.alertService.success('Logged in! :)');
+                // this.toastService.success('Logged in! :)');
                 this.setInfo();
                 if (this.authService.isLoggedIn) {
                     localStorage.setItem('username', res.user.username);
